+++ tests/Functional/ControllerTest.php
@@ -178,6 +178,15 @@ final class ControllerTest extends WebTestCase
                 ];
             }
         }
+
+        if (version_compare(Kernel::VERSION, '7.1.0', '>=')) {
+            yield 'Symfony 7.1 MapUploadedFile attribute' => [
+                [
+                    'name' => 'MapUploadedFileController',
+                    'type' => $type,
+                ],
+            ];
+        }
     }
 
     public static function provideAnnotationTestCases(): \Generator