+++ CHANGELOG.md
@@ -46,6 +46,7 @@ nelmio_api_doc:
 ```
 
 ## 4.34.0
+* Added support for the `#[MapUploadedFile]` symfony controller argument attribute
 * Changed minimum Symfony version for 7.x from 7.0 to 7.1
 
 ## 4.33.6